src/requirements.txt
requests==2.32.2
python-dotenv==1.0.1
keyring==25.2.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.130.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
flask==3.0.3
flask-cors==4.0.1
cachetools==5.3.3
pytz==2024.1
pytest==8.2.1
requests-mock==1.12.1
pytest-mock==3.14.0
